updated new service call changes

+++ src/app/dashboard/my-cart/cart-product/cart-product.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import {ProductService} from '../../../shared';
 import {AppService} from '../../../app.service';
-
+declare var $: any;
 @Component({
   selector: 'app-cart-product',
   templateUrl: './cart-product.component.html',
@@ -27,7 +27,24 @@ export class CartProductComponent implements OnInit {
 		this.productService.changeQuanity(this.product.id,this.product.quantity);
 	  }
   }
-  removeCartItem(id){
-	 this.productService.removeCartItem(this.product.id); 
+  removeCartItem(){
+	  this.productService.removeCartItemService(this.product.id)
+		.subscribe((data: any) => {
+			if(data){
+				this.productService.removeCartItem(this.product.id);
+				this.appService.onShowPreloader.emit(false);
+				$.notify(this.product.title + " has been successfully removed from cart.",'success');
+			} else {
+				if(data.message){
+					$.notify(data.message,'error');
+				} else {
+					$.notify('Product removing from cart failed due to an error. Try after some time.',"error");
+				}
+				this.appService.onShowPreloader.emit(false);
+			}
+		},(data: any) => {
+			this.appService.onShowPreloader.emit(false);
+			$.notify('Product removing from cart failed due to an error. Try after some time.',"error");
+		});		 
   }
 }
